pyproject.toml
[project]
name = "fracta"
version = "0.0.0"
description = "Python based web-app for Viewing & Manipulating Timeseries Data."
readme = { file = "README.md", content-type = "text/markdown" }
authors = [{ name = "jack_of_some_trades" }]
license = { text = "TBD" }
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.39.1",
    "dotenv>=0.9.9",
    "pandas>=2.2.2",
    "pandas-market-calendars>=4.6.0",
    "pywebview>=5.1",
]

[tool.setuptools]
packages = ["fracta", "fracta.orm", "fracta.indicators", "fracta.broker_apis"]
include-package-data = true

[tool.setuptools.package-data]
fracta = ["frontend/*", "indicators/*", "broker_apis/*", "orm/*"]

[project.optional-dependencies]
dist = [
    "setuptools>=80.7.1",
    "twine>=6.1.0",
    "wheel>=0.45.1",
    "build>=1.2.2.post1",
]

[tool.black]
line-length = 120

[tool.pylint."MESSAGES CONTROL"]
max-line-length=120

[project.entry-points."fracta.indicator_pkg"]
pkg_info = "fracta.indicators.__metadata__:PKG_INFO"
